style: #60 코드 검사 페이지 검사 모달 및 toast 위치 수정, 코드 영역 크기 수정
